requirements.txt
GitPython==3.1.12
loguru==0.5.3
semver==2.13.0
